Move to double precision
Move the entities to double precision. The entities will be rendered using x and y co-ordinates in double precision. Also changed speed to use doubles which will lead into new work with calculating ball collision, angle and speed. This will also help with potential balancing and range of speeds allowed as we now have a larger range to experiment with.

+++ src/main/java/com/pong/model/entity/Computer.java
@@ -22,8 +22,8 @@ public class Computer extends Entity {
 
     public final StateMachine<Computer, ComputerState> stateMachine = new AIStateMachine<Computer, ComputerState>(this);
 
-    private final int normalMoveSpeed;
-    private final int idleMoveSpeed = 1;
+    private final double normalMoveSpeed;
+    private final double idleMoveSpeed = 1;
     private final double sightRange;
     private Direction idleMoveDirection;
     private long idleMoveTime = 0;
@@ -53,7 +53,7 @@ public class Computer extends Entity {
      */
     public void moveAndTrackBall() {
         setBaseSpeed(normalMoveSpeed);
-        int ballY = pongModel.getBall().getY();
+        double ballY = pongModel.getBall().getY();
 
         // check if within top bounds
         // and the ball y is less than the paddle y
@@ -116,7 +116,7 @@ public class Computer extends Entity {
     /**
      * {@inheritDoc}
      */
-    public int getSpeed() {
+    public double getSpeed() {
         return baseSpeed + modifiedSpeed;
     }
